[Presentation] Move some utils
